+++ src/Buy2.Application/Common/Interfaces/IUnitOfWork.cs
@@ -1,6 +1,9 @@
-﻿namespace Buy2.Application.Common.Interfaces;
+namespace Buy2.Application.Common.Interfaces;
 
 public interface IUnitOfWork
 {
     Task<int> SaveChangesAsync(CancellationToken cancellationToken = default);
+    Task BeginTransactionAsync(CancellationToken cancellationToken = default);
+    Task CommitTransactionAsync(CancellationToken cancellationToken = default);
+    Task RollbackTransactionAsync(CancellationToken cancellationToken = default);
 }